refactor: create generateArticleMetadata
